configure blueprint
