+++ label_studio/tests/conftest.py
@@ -1,0 +1,1 @@
+from .base import label_studio_app, test_client, captured_templates
